+++ telefunc/plugin/vite/transform.ts
@@ -1,7 +1,6 @@
-import { init, parse } from "es-module-lexer";
-import { relative } from "path";
 import { Plugin } from "vite";
-import { assert, isObject } from "../server/utils";
+import { assert, isObject } from "../../server/utils";
+import {transformTelefuncFile} from "../transformTelefuncFile";
 
 export { transform };
 
@@ -22,27 +21,10 @@ function transform(): Plugin {
       }
       if (id.includes(".telefunc.")) {
         assert(root);
-        const filePath = "/" + relative(root, id);
-        assert(!filePath.startsWith("/."));
-        await init;
-        const exports = parse(src)[1];
-        return {
-          code: getCode(exports, filePath),
-          map: null,
-        };
+        return transformTelefuncFile(src, id, root)
       }
     },
   };
-}
-
-function getCode(exports: readonly string[], filePath: string) {
-  let code = `import { server } from 'telefunc/client';
-
-`;
-  exports.forEach((exportName) => {
-    code += `export const ${exportName} = server['${filePath}:${exportName}'];\n`;
-  });
-  return code;
 }
 
 function isSSR(options: undefined | boolean | { ssr: boolean }): boolean {
